Convert index.php into config file and move server to class file
Fixes #2: index.php is now considered a configuration file (which loads
the server at the end). To avoid conflicts during update the index.php
is not committed to repository, instead a index.php.example is provided.
It is also possible to move index.php somewhere outside the document
root and include it from the real index.php (see instructions in the
index.php.example).

The whole project is now a Composer package. Composer is not neccessary
when running this as standalone server, but it allows to use this
project as a library.

Also, it fixes #6 - a missing parameter in checking for duplicates.